+++ src/app/Sidebar/components/loader.ts
@@ -1,6 +1,3 @@
-import lottieFiles from "../constants/lottieFiles";
-import getRandomLottieFile from "../utils/getRandomLottieFile";
-
 const loader = () => {
   const skeleton = `<div class="flex animate-pulse items-center gap-4">
     <div class="size-10 rounded-md bg-slate-300"></div>
@@ -17,26 +14,16 @@ const loader = () => {
         </div>`;
 
   const skeletons = Array.from({ length: 20 }, () => skeleton).join("");
-  const activeLottieFileId = getRandomLottieFile().id;
-  const lottieAnimations = lottieFiles
-    .map(
-      (lottieFile) => `<div id="lottie-animation-${
-        lottieFile.id
-      }" class="lottieAnimation relative z-10 size-full px-6 py-4 ${
-        lottieFile.id === activeLottieFileId ? "" : "hidden"
-      }">
-        <dotlottie-player id="loader-dotlottie-player" src="${
-          lottieFile.url
-        }" background="transparent" speed="1" loop autoplay class="size-full"></dotlottie-player>
-      </div>`
-    )
-    .join("");
+  const lottieAnimation = `
+      <div id="lottie-animation-10" class="lottieAnimation relative z-10 size-full px-6 py-4">
+        <dotlottie-player id="loader-dotlottie-player" src="https://lottie.host/3ad18b31-d21a-4126-a571-fba2a51ca606/kizxZU1mun.lottie" background="transparent" speed="1" loop autoplay class="size-full"></dotlottie-player>
+      </div>`;
 
   return `<div id="loader" class="hidden relative h-screen w-full overflow-hidden">
       <div class="absolute inset-0 z-0 space-y-4 blur-sm mt-20 p-4 size-full">
         ${skeletons}
       </div>
-      ${lottieAnimations}
+      ${lottieAnimation}
     </div>`;
 };