+++ src/app/Sidebar/sidebar.ts
@@ -6,7 +6,6 @@ import removeBorder from "./utils/border/removeBorder";
 import resetHover from "./utils/border/resetHover";
 import updateBorder from "./utils/border/updateBorder";
 import createDocumentationItem from "./utils/createDocumentationItem";
-import getRandomLottieFile from "./utils/getRandomLottieFile";
 import sortDocumentations from "./utils/sortDocumentations";
 import updateHover from "./utils/updateHover";
 
@@ -221,7 +220,8 @@ const setupEventListeners = () => {
     } else {
       actionItem.classList.remove(
         "hover:cursor-pointer",
-        "hover:brightness-90"
+        "hover:brightness-90",
+        "brightness-50"
       );
     }
   });
@@ -262,15 +262,7 @@ const setupEventListeners = () => {
         documentationList.style.setProperty("display", "none");
         loader.style.setProperty("display", "flex");
 
-        const activeLottieFileId =
-          "lottie-animation-" + getRandomLottieFile().id.toString();
-        lottieAnimations.forEach((lottieAnimation) => {
-          if (lottieAnimation.id === activeLottieFileId) {
-            lottieAnimation.classList.remove("hidden");
-          } else {
-            lottieAnimation.classList.add("hidden");
-          }
-        });
+        hideRegistries = [];
 
         vscode.postMessage({
           type: "searchDocumentation",
@@ -286,15 +278,7 @@ const setupEventListeners = () => {
       documentationList.style.setProperty("display", "none");
       loader.style.setProperty("display", "flex");
 
-      const activeLottieFileId =
-        "lottie-animation-" + getRandomLottieFile().id.toString();
-      lottieAnimations.forEach((lottieAnimation) => {
-        if (lottieAnimation.id === activeLottieFileId) {
-          lottieAnimation.classList.remove("hidden");
-        } else {
-          lottieAnimation.classList.add("hidden");
-        }
-      });
+      hideRegistries = [];
 
       vscode.postMessage({
         type: "searchDocumentation",