+++ src/lib/utils/documentation/searchDocumentation.ts
@@ -1,10 +1,11 @@
 import IDependency from "../../interfaces/IDependency";
 import IDocumentation from "../../interfaces/IDocumentation";
+import IPackageInformation from "../../interfaces/IPackageInformation";
 import checkIframeSupport from "../checkIframeSupport";
 import findUrlDocumentation from "../findUrlDocumentation";
 import getFaviconUrl from "../getFaviconUrl";
 import { DocumentationViewProvider } from "../provider/DocumentationViewProvider";
-import searchPackage from "../searchPackage";
+import searchPackage from "../searchPackage/searchPackage";
 import formatUrl from "./formatUrl";
 
 const searchDocumentation = async (
@@ -18,7 +19,7 @@ const searchDocumentation = async (
     const uniqueIds: string[] = [];
 
     const documentations = await Promise.all(
-      packages.map(async (pkg) => {
+      packages.map(async (pkg: IPackageInformation) => {
         const info = pkg;
 
         const registry = info.registry;